Enough to push it to main
I still have a lot of work to do here. But I wanted to get it onto main and let people see it finally. It still very much looks like a mobile version but it is responsive and should be easily viewable on all devices for at least chrome, edge, & firefox. It needs accessibility, and compatability with IE and some other browsers but that can come in time. The desktop version needs to be done which is going to be way more robust than the mobile version your seeing now.

* Completed enough to let it be viewed
* Made some global classes to help edit easier
* Added media queries at the bottom to responsively define text and keep the look if you view on PC even though its coded for mobile at the moment.

* Made the logo clickable to get back to the home page

* Added more content on the mission page